(refactor): normalize app bootstrap into core/features/surfaces
